Get rid of `MOVSX` instructions when adding a pointer with an `int`.

+++ MCFCRT/src/ext/rawmemchr.c
@@ -14,7 +14,7 @@ void *_MCFCRT_rawmemchr(const void *s, int c){
 	rp = (const char *)((uintptr_t)rp & (uintptr_t)-32);
 	__m128i xc[1];
 	__MCFCRT_xmmsetb(xc, (uint8_t)c);
-	int shift = (int)((const char *)s - rp);
+	unsigned shift = (unsigned)((const char *)s - rp);
 	uint32_t skip = (uint32_t)-1 << shift;
 	for(;;){
 		__m128i xw[2];
@@ -22,7 +22,7 @@ void *_MCFCRT_rawmemchr(const void *s, int c){
 		__MCFCRT_xmmload_2(xw, rp, _mm_load_si128);
 		mask = __MCFCRT_xmmcmp_21b(xw, xc, _mm_cmpeq_epi8) & skip;
 		if(_MCFCRT_EXPECT_NOT(mask != 0)){
-			shift = __builtin_ctzl(mask);
+			shift = (unsigned)__builtin_ctzl(mask);
 			return (char *)rp + shift;
 		}
 		rp += 32;
