+++ MCFCRT/src/stdc/wchar/wcschr.c
@@ -18,7 +18,7 @@ wchar_t *wcschr(const wchar_t *s, wchar_t c){
 	__MCFCRT_xmmsetz(xz);
 	__m128i xc[1];
 	__MCFCRT_xmmsetw(xc, (uint16_t)c);
-	int shift = (int)((const wchar_t *)s - rp);
+	unsigned shift = (unsigned)((const wchar_t *)s - rp);
 	uint32_t skip = (uint32_t)-1 << shift;
 	for(;;){
 		__m128i xw[4];
@@ -36,7 +36,7 @@ wchar_t *wcschr(const wchar_t *s, wchar_t c){
 		mask = __MCFCRT_xmmcmp_41w(xw, xc, _mm_cmpeq_epi16) & skip;
 		if(_MCFCRT_EXPECT_NOT(mask != 0)){
 found:
-			shift = __builtin_ctzl(mask);
+			shift = (unsigned)__builtin_ctzl(mask);
 			return (wchar_t *)rp + shift;
 		}
 		rp += 32;
